Refactor: Update database connection and remove unused states
