+++ pages/api/admin/promos/index.js
@@ -1,6 +1,6 @@
-import { connectDB } from '@/libs/mongodb';
+import connectMongo from '@/libs/mongoose';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/pages/api/auth/[...nextauth]';
+import { authOptions } from '@/libs/next-auth';
 import Promo from '@/models/Promo';
 
 export default async function handler(req, res) {
@@ -18,7 +18,7 @@ export default async function handler(req, res) {
     return res.status(403).json({ error: 'Forbidden - Admin access required' });
   }
 
-  await connectDB();
+  await connectMongo();
 
   switch (req.method) {
     case 'GET':
